+++ doc/whatsnew.rst
@@ -10,8 +10,9 @@ v0.2.2
 * Make HTTP connections configurable by exposing the 
   `requests.get API <http://www.python-requests.org/en/latest/>`_ 
   through the :class:`pandasdmx.api.Request` constructor.
-  Hence, proxy servers, authentication information and other HTTP-related parameters consumed by ``requests.get`` can be
-  specified for each ``Request`` instance and used in subsequent requests.
+  Hence, proxy servers, authorisation information and other HTTP-related parameters consumed by ``requests.get`` can be
+  specified for each ``Request`` instance and used in subsequent requests. The configuration is exposed as a dict through
+  a new ``Request.client.config`` attribute.
 * Responses have a new ``http_headers`` attribute containing the HTTP headers returned by the SDMX server
 
 v0.2.1